kernel: Refactor square_dist_scalar_mult
The 'square_dist_scalar_mult' kernel could be very useful for symbol
demapping. Currently, the AVX implementation seems to be slower than the
generic one. Also, the code is hard to read and needs refactoring.
Especially, move common intrinsics to `include/*_intrinsics.h` in order
to have a simple way to implement aligned and unaligned kernels.