Py3k: Add comment for CMake 3.12+
With this commit I added a comment that CMake 3.12+ supports a simpler
way to find Python and prefer Py3k over Py2k. Especially, it
automatically prefers Py3k as desired.
Since we set CMake 3.8 as our minimum required version, I leave this in
as a `FUTURE TODO` as soon as we move to a newer minimum CMake version.