Improve performance of generic kernels with complex multiply
This PR incorporates the cx-limited-range checks for complex
multiply similar to GR pull request 3114.  Native complex
multiply math has a significant amount of processing overhead
ensuring that INF and NAN conditions are not encountered.
In some tests on a modern processor, these extra tests cause
a simle c = a * b complex calculation to take as much as 20%
of the processing time in a loop.  The -fcx-limited-range
compiler flag, when available ensures that the c = a * b
calculation is reduced to just the appropriate math.  The
changes to the makefile ensure that the compiler supports it
before enabling it.  On embedded debian-based platforms where
the generic kernel may get chosen, this should have a significant
positive impact performance for complex multiply kernels.