rotator: Refactor AVX kernels
Fix #331

This is ongoing work to track down the root cause of #329.

1. Make sure all input scalars are normalized -> no improvement
2. Trigger renormalization after same num_points -> no improvement

Lower tolerance values move the first point that breaks this bound
further to the front. It is not obvious why this would happen so far.

Since SSE kernels tend to break later, approx. factor 2, maybe the
generic kernels accumulates errors faster. That's just a wild guess
though.