ci: Add Github CI Action
This commit adds a CI tests for GitHub Actions for Linux, Windows and
MacOS.

Since we use GitHub for code hosting, it is worthwhile to have CI tests
that run in every VOLK repo on push. Be aware that Windows and MacOS CI
does only run generic kernels at the moment because it cannot find
`xgetbv`.