ci: Remove Ubuntu 16.04 GCC5 test on TravisCI
This particular combination of Ubuntu and GCC does not compile due
to some bug in this Ubuntu version. Since this is a really old
Ubuntu version we decided to drop this test.

Users who want to use VOLK with this distro are adviced to stick with
older versions. It would be too much of a burden to work around a bug in
outdated software.