readme: Add section on supported platforms
So far, this is kind of a work in progress. I'd like to discuss the
exact wording here.
We run CI on Ubuntu, Windows and MacOS. For Windows and MacOS, we focus
on the latest release and do not make any attempt to support older
releases.
For Linux, we want to have support for distros other than Ubuntu and
older distros to some degree. Though, we decided to drop support for
Ubuntu 16.04 because some bugs make it hard work put a work around in
place for this old distro.