contributing: Add CONTRIBUTING.md and DCO.txt
It is recommended good practice to have a CONTRIBUTING.md file. I used
the GNU Radio file as a template and adopted and simplified it a bit.
Since GNU Radio switch to DCO, I assume it is a welcome improvement for
VOLK as well.

I added a DCO.txt as well. I suggest we move from a CLA based
contribution system to a DCO based one. I hope this decreases the entry
hurdle to new contributors.

Just like GNU Radio I'd like to use [DCObot][https://probot.github.io/apps/dco/]
for VOLK as well.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>