ci: Add more CI to GitHub Actions
We see issues with TravisCI because of their policy change. Here, I
start to move tests to Github Actions. Particularly the default builds
with clang and GCC are now on GH Actions.

Apparently, Actions does not yet support Clang 11. Also, old compilers
are left out. Anything before GCC/Clang 7 is excluded. For GCC, CMake
reports issues and some packages seem to be missing. For Clang, packages
for these compilers are missing.

Actions does not support ARM etc. Thus, only x86 is supported. We might
be able to add QEMU + cross-compiles. Also, ARM support seems to be on
their roadmap for Q1'21.

Another step is included in the CI chain that prints the current
configuration and hardware info as found by `list_cpu_features` and
`volk-config-info`.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>