lgpl: add Andy to the list
We received the following statement:

I, George Andrew Walls, hereby resubmit all my
contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later.

My GitHub handle is awalls-cx18.
My email addresses used for contributions are:
        andy@silverblocksystems.net
        awalls.cx18@gmail.com
        awalls@md.metrocast.net
.

I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

Signed-off-by: Michael Dickens <michael.dickens@ettus.com>