lgpl: Add Stefan to the list
We received the following statement

I, Stefan Oltmanns, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is Stefan-Olt. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/a37249ec95df5387bf327b4a84910c03ec671a65#commitcomment-58914315

Signed-off-by: Johannes Demel <demel@uni-bremen.de>