lgpl: Add Julien to the list
We received the following statement

I, Julien Olivain, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle jolivain. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/cc1abd2ea27c7f303efaa6466e979700000915ed#commitcomment-59558033

Signed-off-by: Johannes Demel <demel@uni-bremen.de>