lgpl: Add Adam to the list
We received approval for the following statement

I, Adam Thompson, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle awthomp. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/ea2f0b87f9c6650a3c42a279b903e4d049cc8640#commitcomment-58894646

Signed-off-by: Johannes Demel <demel@uni-bremen.de>