lgpl: Add Roy to the list
We received approval for the following statement

I, Roy Thompson, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle `roythompson`.  I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/8464c3b8ffbdb739f2a41c65f506a5a6c7f4bb60#commitcomment-58923065

Signed-off-by: Johannes Demel <demel@uni-bremen.de>