lgpl: Add luzpaz to the list
We received approval for the following statement

I, luzpaz, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is luzpaz.  I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/1e2aeb597eaa80a9a7481ecc2b65875f9118d997#commitcomment-58927395

Signed-off-by: Johannes Demel <demel@uni-bremen.de>