lgpl: Add Damian to the list
We received approval for the following statement

I, Damian Miralles, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is `dmiralles2009`. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/bc21d2d519ebffd2809068592dcd6382f5d12a8a#commitcomment-58938627

Signed-off-by: Johannes Demel <demel@uni-bremen.de>