cpu_features: Update submodule pointer
We have 3 issues that should be fixed with this commit.

Fix #428
Should be fixed because cpu_features detects `arm64` now. Thus, it
builds on MacOS and reports M1 capabilities.

Fix #478
Fix #484
cpu_features received quite a bit of contributions for FreeBSD. All the
issues we had should be fixed now. However, this might require further
evaluation.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>